+++ src/main/java/net/jptrzy/trinkets/curios/theme/integrations/ScoutUtils.java
@@ -10,7 +10,7 @@ public abstract class ScoutUtils {
     public static boolean haveLeftSatchel(TrinketComponent trinkets){
         for (int i=0; i<trinkets.getAllEquipped().size(); i++) {
             if (trinkets.getAllEquipped().get(i).getRight().getItem() instanceof BaseBagItem &&
-                    trinkets.getAllEquipped().get(i).getLeft().index() == 0 &&
+                    //trinkets.getAllEquipped().get(i).getLeft().index() == 0 &&
                     Objects.equals(trinkets.getAllEquipped().get(i).getLeft().inventory().getSlotType().getName(), "pouch")) {
                 return true;
             }